requirements.txt
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
flask>=2.0.0
flask-cors>=3.0.0
Pillow>=9.0.0

